+++ src/main/java/com/jetblue/api/validator/LongitudeConstraintValidator.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Component;
  * The Class LatitudeConstraintValidator.
  */
 @Component
-public class LongitudeConstraintValidator implements ConstraintValidator<ValidLongitude, Long> {
+public class LongitudeConstraintValidator implements ConstraintValidator<ValidLongitude, Double> {
 
 	private static final String LONGITUDE_PATTERN="^(\\+|-)?(?:180(?:(?:\\.0{1,6})?)|(?:[0-9]|[1-9][0-9]|1[0-7][0-9])(?:(?:\\.[0-9]{1,6})?))$";
 	/*
@@ -33,12 +33,12 @@ public class LongitudeConstraintValidator implements ConstraintValidator<ValidLo
 	 * javax.validation.ConstraintValidatorContext)
 	 */
 	@Override
-	public boolean isValid(final Long longitude, final ConstraintValidatorContext context) {
+	public boolean isValid(final Double longitude, final ConstraintValidatorContext context) {
 		if(null == longitude) {
 			return false;
 		}
 		Pattern longitudePattern = Pattern.compile(LONGITUDE_PATTERN);
-		Matcher matcher = longitudePattern.matcher(Long.toString(longitude));
+		Matcher matcher = longitudePattern.matcher(Double.toString(longitude));
 		return matcher.matches();
 	}
